// cypress/integration/ascend/interfaces/ReadinessAssessmentTypes.ts
export enum NewAssessmentLabels {
    NewEngagement = 'New Engagement',
    ReadinessAssessment = 'Readiness Assessment',
    ShortNameField = 'Short Name',
    ClientField = 'Client',
    ClientString = 'Client string',
    ShortNameString = 'Short Name string',
    ContactField = 'Contact',
    ContactEmail = 'Contact email',
    ClientLogoField = 'Client Logo',
    ClientLogoInput = 'Client Logo Input',
    ClientLogoFile = 'Client Logo file',
    AuditorAssistedField = 'Auditor Assisted',
    AuditorAssistedBoolean = 'Auditor Assisted boolean',
    ExistingClientLogo = 'existing client logo',
    EditPencilIcon = 'edit pencil',
    DragAndDrop = 'Drag and drop here or browse to upload',
    AssessmentField = 'Assessment',
    AssessmentString = 'Assessment string',
    YearField = 'Year',
    YearString = 'Year string',
    AssessmentNameField = 'Assessment Name',
    AssessmentNameString = 'Assessment Name string',
    VersionField = 'Version',
    VersionString = 'Version string',
    CategoriesField = 'Categories',
    CategoriesString = 'Categories string',
    ExecutiveField = 'Executive',
    ExecutiveString = 'Executive string',
    PracticeLeadField = 'Practice Lead',
    PracticeLeadString = 'Practice Lead string',
    ManagerField = 'Manager',
    ManagerString = 'Manager string',
    ConsultantField = 'Consultant',
    ConsultantString = 'Consultant string',
    Publish = 'Publish',
    Snackbar = 'Snackbar',
    SearchBar = 'Search bar',
    Manage = 'Manage',
    Continue = 'Continue',
}